Refactor elements to be a class
Rather than using an array where the first element is the name of the
tag, use an Element class that contains the name, attribute and children
as separate properties.  This is easier to reason with and allows arrays
along with all other iterable objects to be unpacked automatically which
makes compositions easier to express in code.